requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
jinja2==3.1.2
reportlab==4.0.7
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0